moving all containerd tests one file and fix profile
